Remove the build script for appveyor
During measurement-kit/mkmmdb#2, it become obvious that the way
in which I was running AppVeyor wasn't correct.

The builds were not failed even though they out to.

This commit removes the build script for AppVeyor and I'll revert
the code to run the commands directly from the .yml file.

After all, they're literally three lines of code.